requirements.txt
accelerate==1.10.1
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.0rc6
idna==3.11
jinja2==3.1.6
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==12.0.0
psutil==7.1.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pyyaml==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@bf0bce8d5fcbabcde86e63c2634e87ef8c2614dc
typer-slim==0.19.2
typing-extensions==4.15.0
urllib3==2.5.0
